Indent the document for legibility
We use `xmllint --format` to apply reliable indentation
and formatting to the document.

Nokogiri doesn't apply indents to the merged <trkpts> despite
the use of:

`doc.to_xml(indent: 2, encoding: "utf-8")`

So we add an additional dependency. `xmllint` can be installed
as a part of `libxml2-utils` with:

```
apt install libxml2-utils
```